requirements.txt
pyspark==3.4.1
pandas
numpy
pyarrow
requests
azure-storage-blob
matplotlib
jupyter
ollama
pytest